astype can not be generic

+++ spynnaker/spynnaker_plotting.py
@@ -20,6 +20,7 @@ https://github.com/NeuralEnsemble/PyNN/blob/master/pyNN/utility/plotting.py
 from typing import Any, Dict, Final, List, Union
 from types import ModuleType
 
+import numpy
 from neo import SpikeTrain, Block, Segment, AnalogSignal
 from neo.core.spiketrainlist import SpikeTrainList  # type: ignore[import]
 import numpy as np
@@ -160,8 +161,8 @@ def heat_plot_numpy(axes: Axes, data: NDArray, label: str = '',
     :param str label: Label for the graph
     :param options: plotting options
     """
-    neurons = data[:, 0].astype(int)
-    times = data[:, 1].astype(int)
+    neurons = data[:, 0].astype(numpy.uint32)
+    times = data[:, 1].astype(numpy.uint32)
     values = data[:, 2]
     info_array = np.empty((max(neurons) + 1, max(times) + 1))
     info_array[:] = np.nan
